[bucket C] fix: dedupe _CustomModelFixture test doubles into shared conftest
tests/task_test.py, custom/tests/assembler_test.py, and
_private/model_class/tests/resolve_test.py each defined an identical
minimal Model subclass used only as an importable dotted path for
framework-resolution tests. Move the single definition into a package-
root conftest.py (collected by pytest for all three test packages) and
update the dotted-path constants to point at it.
