+++ python/michelangelo/workflow/tasks/tabular_assembler/_private/model_class/tests/resolve_test.py
@@ -3,12 +3,10 @@
 from __future__ import annotations
 
 import unittest
-from typing import TYPE_CHECKING
 
 import pytorch_lightning as pl
 import torch.nn as nn
 
-from michelangelo.lib.model_manager.interface.custom_model import Model
 from michelangelo.workflow.tasks.tabular_assembler._private.model_class.resolve import (
     resolve_model_class,
     resolve_training_framework,
@@ -20,27 +18,8 @@ from michelangelo.workflow.variables.metadata import (
     TRAINING_FRAMEWORK_PYTORCH,
 )
 
-if TYPE_CHECKING:
-    from numpy import ndarray
-
-
-class _CustomModelFixture(Model):
-    """Minimal concrete ``Model`` used only to exercise framework resolution."""
-
-    def save(self, path: str) -> None:
-        pass
-
-    @classmethod
-    def load(cls, path: str) -> _CustomModelFixture:
-        return cls()
-
-    def predict(self, inputs: dict[str, ndarray]) -> dict[str, ndarray]:
-        return inputs
-
-
 _CUSTOM_MODEL_PATH = (
-    "michelangelo.workflow.tasks.tabular_assembler._private.model_class.tests."
-    "resolve_test._CustomModelFixture"
+    "michelangelo.workflow.tasks.tabular_assembler.conftest._CustomModelFixture"
 )
